pyproject.toml
[tool.poetry]
name = "resume-ai"
version = "0.1.0"
description = ""
authors = ["Kerim Tricic <kerim.tricic@dualsharks.com>"]
readme = "README.md"
package-mode = false
[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.46.1"
langchain = "^0.3.26"
langchain-community = "^0.3.26"
chromadb = "^1.0.15"
openai = "^1.93.0"
tiktoken = "^0.9.0"
docling = "^2.39.0"
protobuf = "3.20.3"
python-dotenv = "^1.0.0"
pypdf = "^5.7.0"
faiss-cpu = "^1.11.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"
pandas = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
